fix cau truc
