+++ src/main/java/sorfware/example/sorfware/model/dto/request/SearchRequest.java
@@ -1,0 +1,16 @@
+package sorfware.example.sorfware.model.dto.request;
+
+
+
+public class SearchRequest {
+    private String keyword;
+
+    public String getKeyword() {
+        return keyword;
+    }
+
+    public void setKeyword(String keyword) {
+        this.keyword = keyword;
+    }
+}
+
