+++ src/main/java/sorfware/example/sorfware/model/dto/response/SearchResult.java
@@ -1,17 +1,29 @@
 package sorfware.example.sorfware.model.dto.response;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import sorfware.example.sorfware.model.entity.Message;
-import sorfware.example.sorfware.model.entity.User;
-
 import java.util.List;
 
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
 public class SearchResult {
-    private List<User> users;
-    private List<Message> messages;
+    private List<?> users;
+    private List<?> messages;
+
+    public SearchResult(List<?> users, List<?> messages) {
+        this.users = users;
+        this.messages = messages;
+    }
+
+    public List<?> getUsers() {
+        return users;
+    }
+
+    public void setUsers(List<?> users) {
+        this.users = users;
+    }
+
+    public List<?> getMessages() {
+        return messages;
+    }
+
+    public void setMessages(List<?> messages) {
+        this.messages = messages;
+    }
 }
